Revert "internals: Server rendering"
