+++ app/app.js
@@ -10,10 +10,12 @@ import 'babel-polyfill';
 
 // Import all the third party stuff
 import React from 'react';
+import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { ConnectedRouter } from 'react-router-redux';
-
-import PropTypes from 'prop-types';
+import FontFaceObserver from 'fontfaceobserver';
+import createHistory from 'history/createBrowserHistory';
+import 'sanitize.css/sanitize.css';
 
 // Import root app
 import App from 'containers/App';
@@ -21,23 +23,92 @@ import App from 'containers/App';
 // Import Language Provider
 import LanguageProvider from 'containers/LanguageProvider';
 
+// Load the favicon, the manifest.json file and the .htaccess file
+/* eslint-disable import/no-webpack-loader-syntax */
+import '!file-loader?name=[name].[ext]!./images/favicon.ico';
+import '!file-loader?name=[name].[ext]!./images/icon-72x72.png';
+import '!file-loader?name=[name].[ext]!./images/icon-96x96.png';
+import '!file-loader?name=[name].[ext]!./images/icon-120x120.png';
+import '!file-loader?name=[name].[ext]!./images/icon-128x128.png';
+import '!file-loader?name=[name].[ext]!./images/icon-144x144.png';
+import '!file-loader?name=[name].[ext]!./images/icon-152x152.png';
+import '!file-loader?name=[name].[ext]!./images/icon-167x167.png';
+import '!file-loader?name=[name].[ext]!./images/icon-180x180.png';
+import '!file-loader?name=[name].[ext]!./images/icon-192x192.png';
+import '!file-loader?name=[name].[ext]!./images/icon-384x384.png';
+import '!file-loader?name=[name].[ext]!./images/icon-512x512.png';
+import '!file-loader?name=[name].[ext]!./manifest.json';
+import 'file-loader?name=[name].[ext]!./.htaccess'; // eslint-disable-line import/extensions
+/* eslint-enable import/no-webpack-loader-syntax */
+
+import configureStore from './configureStore';
+
+// Import i18n messages
+import { translationMessages } from './i18n';
+
 // Import CSS reset and Global Styles
 import './global-styles';
 
-const Root = ({ messages, history, store }) => (
-  <Provider store={store}>
-    <LanguageProvider messages={messages}>
-      <ConnectedRouter history={history}>
-        <App />
-      </ConnectedRouter>
-    </LanguageProvider>
-  </Provider>
-);
+// Observe loading of Open Sans (to remove open sans, remove the <link> tag in
+// the index.html file and this observer)
+const openSansObserver = new FontFaceObserver('Open Sans', {});
 
-Root.propTypes = {
-  messages: PropTypes.object.isRequired,
-  history: PropTypes.object.isRequired,
-  store: PropTypes.object.isRequired,
+// When Open Sans is loaded, add a font-family using Open Sans to the body
+openSansObserver.load().then(() => {
+  document.body.classList.add('fontLoaded');
+}, () => {
+  document.body.classList.remove('fontLoaded');
+});
+
+// Create redux store with history
+const initialState = {};
+const history = createHistory();
+const store = configureStore(initialState, history);
+const MOUNT_NODE = document.getElementById('app');
+
+const render = (messages) => {
+  ReactDOM.render(
+    <Provider store={store}>
+      <LanguageProvider messages={messages}>
+        <ConnectedRouter history={history}>
+          <App />
+        </ConnectedRouter>
+      </LanguageProvider>
+    </Provider>,
+    MOUNT_NODE
+  );
 };
 
-export default Root;
+if (module.hot) {
+  // Hot reloadable React components and translation json files
+  // modules.hot.accept does not accept dynamic dependencies,
+  // have to be constants at compile-time
+  module.hot.accept(['./i18n', 'containers/App'], () => {
+    ReactDOM.unmountComponentAtNode(MOUNT_NODE);
+    render(translationMessages);
+  });
+}
+
+// Chunked polyfill for browsers without Intl support
+if (!window.Intl) {
+  (new Promise((resolve) => {
+    resolve(import('intl'));
+  }))
+    .then(() => Promise.all([
+      import('intl/locale-data/jsonp/en.js'),
+      import('intl/locale-data/jsonp/de.js'),
+    ]))
+    .then(() => render(translationMessages))
+    .catch((err) => {
+      throw err;
+    });
+} else {
+  render(translationMessages);
+}
+
+// Install ServiceWorker and AppCache in the end since
+// it's not most important operation and if main code fails,
+// we do not want it installed
+if (process.env.NODE_ENV === 'production') {
+  require('offline-plugin/runtime').install(); // eslint-disable-line global-require
+}
