+++ app/containers/App/index.js
@@ -132,16 +132,8 @@ NavigationFooterFromRouteDescriptions.propTypes = {
 const RoutableRouteDescriptions = withRouter(RouteDescriptions);
 const NavigationFooterWithRouteDescriptions = withRouteDescriptions(NavigationFooterFromRouteDescriptions);
 
-const MaybeAtlaskitThemeProvider = ({ children }) => (
-  __SERVER__ ? <div>{children}</div> : <AtlaskitThemeProvider>{children}</AtlaskitThemeProvider>
-);
-
-MaybeAtlaskitThemeProvider.propTypes = {
-  children: PropTypes.node.isRequired,
-};
-
 const NavigatablePage = () => (
-  <MaybeAtlaskitThemeProvider>
+  <AtlaskitThemeProvider>
     <ResponsivePage
       title="Intuitive Math"
       sections={sections}
@@ -164,7 +156,7 @@ const NavigatablePage = () => (
         <NavigationFooterWithRouteDescriptions />
       </RoutableRouteDescriptions>
     </ResponsivePage>
-  </MaybeAtlaskitThemeProvider>
+  </AtlaskitThemeProvider>
 );
 
 export default function App() {
