+++ app/containers/HomePage/Loadable.js
@@ -4,9 +4,9 @@
 import React from 'react';
 
 import Loadable from 'react-loadable';
+import Spinner from 'react-spinkit';
 
 import CenteredParagraph from 'components/CenteredParagraph';
-import Spinner from 'components/MonkeyPatchedSpinKit';
 import Section from 'components/Section';
 
 export default Loadable({
