+++ app/index.html
@@ -8,20 +8,20 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
 
     <!-- Allow installing the app to the homescreen -->
-    <link rel="manifest" href="/static/manifest.json">
+    <link rel="manifest" href="/manifest.json">
     <meta name="mobile-web-app-capable" content="yes">
 
     <!-- iOS home screen icons -->
     <meta name="apple-mobile-web-app-title" content="react boilerplate">
-    <link rel="apple-touch-icon" sizes="120x120" href="/static/icon-120x120.png">
-    <link rel="apple-touch-icon" sizes="152x152" href="/static/icon-152x152.png">
-    <link rel="apple-touch-icon" sizes="167x167" href="/static/icon-167x167.png">
-    <link rel="apple-touch-icon" sizes="180x180" href="/static/icon-180x180.png">
+    <link rel="apple-touch-icon" sizes="120x120" href="/icon-120x120.png">
+    <link rel="apple-touch-icon" sizes="152x152" href="/icon-152x152.png">
+    <link rel="apple-touch-icon" sizes="167x167" href="/icon-167x167.png">
+    <link rel="apple-touch-icon" sizes="180x180" href="/icon-180x180.png">
 
     <!-- Google site verification -->
     <meta name="google-site-verification" content="hlRaBQIfmIUqpVumpvejdE7FqMzkHZP5uU5fvnMvbFg" />
 
-    <link rel="icon" href="/static/favicon.ico" />
+    <link rel="icon" href="/favicon.ico" />
     <title>Intuitive Math</title>
   </head>
   <body>
