+++ internals/config.js
@@ -28,7 +28,6 @@ const ReactBoilerplate = {
         'ip',
         'minimist',
         'sanitize.css',
-        'webpack-node-externals',
       ],
 
       /**
