+++ internals/webpack/webpack.base.babel.js
@@ -15,9 +15,7 @@ module.exports = (options) => ({
   entry: options.entry,
   output: Object.assign({ // Compile into js/build.js
     path: path.resolve(process.cwd(), 'build'),
-    publicPath: '/static',
-    libraryTarget: options.libraryTarget,
-    library: options.library,
+    publicPath: '/',
   }, options.output), // Merge with env dependent settings
   module: {
     rules: [
@@ -94,11 +92,9 @@ module.exports = (options) => ({
     // inside your code for any environment checks; UglifyJS will automatically
     // drop any unreachable code.
     new webpack.DefinePlugin({
-      'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV),
-      // Put a define in place if we're server-side rendering
-      ...(options.server ? {
-        __SERVER__: '1',
-      } : {}),
+      'process.env': {
+        NODE_ENV: JSON.stringify(process.env.NODE_ENV),
+      },
     }),
     new webpack.NamedModulesPlugin(),
   ]),
@@ -116,7 +112,6 @@ module.exports = (options) => ({
     ],
   },
   devtool: options.devtool,
-  target: options.target || 'web', // Make web variables accessible to webpack, e.g. window
-  externals: options.externals || [],
+  target: 'web', // Make web variables accessible to webpack, e.g. window
   performance: options.performance || {},
 });
