+++ lambda.js
@@ -1,6 +1,4 @@
 const serverless = require('serverless-http');
 const app = require('./server/server');
 
-module.exports = {
-  handler: serverless(app),
-};
+exports.handler = serverless(app);
