+++ package.json
@@ -21,16 +21,13 @@
     "preinstall": "npm run npmcheckversion",
     "postinstall": "npm run build:dll",
     "prebuild": "npm run build:clean",
-    "build": "npm run build:server && npm run build:client",
-    "build:server": "cross-env NODE_ENV=production webpack --config internals/webpack/webpack.prod.server.babel.js --color -p --progress --hide-modules --display-optimization-bailout",
-    "build:serverless": "cross-env NODE_ENV=production webpack --config internals/webpack/webpack.prod.serverless.babel.js --color --progress --hide-modules --display-optimization-bailout",
-    "build:client": "cross-env NODE_ENV=production webpack --config internals/webpack/webpack.prod.client.babel.js --color -p --progress --hide-modules --display-optimization-bailout",
+    "build": "cross-env NODE_ENV=production webpack --config internals/webpack/webpack.prod.babel.js --color -p --progress --hide-modules --display-optimization-bailout",
     "build:clean": "rimraf ./build",
     "build:dll": "node ./internals/scripts/dependencies.js",
-    "start": "cross-env NODE_ENV=development babel-node server",
+    "start": "cross-env NODE_ENV=development node server",
     "start:tunnel": "cross-env NODE_ENV=development ENABLE_TUNNEL=true node server",
     "start:production": "npm run test && npm run build && npm run start:prod",
-    "start:prod": "cross-env NODE_ENV=production node build/server.js",
+    "start:prod": "cross-env NODE_ENV=production node server",
     "presetup": "npm i chalk shelljs",
     "setup": "node ./internals/scripts/setup.js",
     "postsetup": "npm run build:dll",
@@ -70,9 +67,7 @@
     "env": {
       "production": {
         "only": [
-          "app",
-          "internals",
-          "server"
+          "app"
         ],
         "plugins": [
           "transform-react-remove-prop-types",
@@ -96,9 +91,6 @@
       "node": true,
       "jest": true,
       "es6": true
-    },
-    "globals": {
-      "__SERVER__": true
     },
     "plugins": [
       "redux-saga",
@@ -217,7 +209,6 @@
     "@atlaskit/navigation": "^27.0.4",
     "@atlaskit/page": "^5.0.5",
     "@atlaskit/theme": "^2.2.9",
-    "babel-cli": "^6.26.0",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-polyfill": "6.23.0",
     "chalk": "1.1.3",
@@ -257,7 +248,6 @@
     "styled-components": "2.0.0",
     "three": "^0.89.0",
     "warning": "3.0.0",
-    "webpack-node-externals": "^1.6.0",
     "whatwg-fetch": "2.0.3"
   },
   "devDependencies": {
