+++ scripts/deploy.sh
@@ -27,8 +27,7 @@ PROGRESS=$!
 
 yarn install
 yarn build:dll
-yarn build:client
-yarn build:serverless
+yarn build
 yarn serverless-deploy
 
 kill $PROGRESS >/dev/null 2>&1
