+++ server/middlewares/addProdMiddlewares.js
@@ -1,9 +1,8 @@
 const path = require('path');
 const express = require('express');
-const ssr = require('../ssr');
 
 module.exports = function addProdMiddlewares(app, options) {
-  const publicPath = options.publicPath || '/static';
+  const publicPath = options.publicPath || '/';
   const outputPath = options.outputPath || path.resolve(process.cwd(), 'build');
 
   // compression middleware compresses your server responses which makes them
@@ -13,5 +12,6 @@ module.exports = function addProdMiddlewares(app, options) {
   // Except that for now it doesn't work on serverless.
   // app.use(compression());
   app.use(publicPath, express.static(outputPath));
-  ssr(app);
+
+  app.get('*', (req, res) => res.sendFile(path.resolve(outputPath, 'index.html')));
 };
